+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * `ServiceCollectionServiceExtensions` in Integration `TextFixture` for easier service mocking with `ReplaceServiceWithSingletonMock`
 
+### Updated
+
+* Removed `faulty` producer assertion in integration tests for better performance
+
 ## [0.15.0] - 06/05/2022
 
 ### Added